requirements.txt
flask==3.0.3
flask-login==0.6.3
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.2.0
numpy==1.24.4
fpdf2==2.7.9
requests==2.31.0
werkzeug==3.0.3